Created PlayerPoiBase and PlayerMainMapPoi
